+++ src/services/authService.ts
@@ -1,5 +1,5 @@
+import { axiosInstance } from '@/lib';
 import { AUTH_ENDPOINTS } from '@/constants';
-import { axiosInstance } from '@/lib/axios';
 
 export const verifyEmail = async (email: string) => {
   const { data } = await axiosInstance.post(AUTH_ENDPOINTS.EMAIL_VERIFY, {
